requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
groq==0.4.2
gTTS==2.5.1
elevenlabs==0.3.1
SpeechRecognition==3.10.1
python-multipart==0.0.9
pydantic==2.6.3
requests==2.31.0
aiofiles==23.2.1
